requirements.txt
wxPython
google-generativeai
requests
pytz
apscheduler
sqlalchemy